Moved saturation vapour pressure calculation into only class where it is called.
